# Conflicts:
#	src/mc/common/Globals.h
#	src/mc/deps/core/file/Path.h
#	src/mc/deps/core/math/Color.h
#	src/mc/deps/core/math/Vec2.h
#	src/mc/deps/core/math/Vec3.h
#	src/mc/deps/core/string/HashedString.h
#	src/mc/deps/core/utility/BinaryStream.h
#	src/mc/deps/ecs/gamerefs_entity/OwnerStorageEntity.h
#	src/mc/deps/ecs/gamerefs_entity/StackResultStorageEntity.h
#	src/mc/deps/json/Value.h
#	src/mc/diagnostics/Diagnostics.h
#	src/mc/entity/components/UserEntityIdentifierComponent.h
#	src/mc/nbt/ByteTag.h
#	src/mc/nbt/CompoundTag.h
#	src/mc/nbt/FloatTag.h
#	src/mc/nbt/Int64Tag.h
#	src/mc/nbt/IntTag.h
#	src/mc/nbt/StringTag.h
#	src/mc/network/NetworkBlockPosition.h
#	src/mc/platform/brstd/copyable_function.h
#	src/mc/server/commands/CommandBlockName.h
#	src/mc/server/commands/CommandParameterData.h
#	src/mc/server/commands/CommandRegistry.h
#	src/mc/server/commands/CommandVersion.h
#	src/mc/server/commands/RelativeFloat.h
#	src/mc/util/MolangScriptArg.h
#	src/mc/world/Container.h
#	src/mc/world/actor/Actor.h
#	src/mc/world/actor/DataItem.h
#	src/mc/world/item/ItemInstance.h
#	src/mc/world/item/ItemStack.h
#	src/mc/world/item/ItemStackBase.h
#	src/mc/world/item/SaveContext.h
#	src/mc/world/level/BlockPos.h
#	src/mc/world/level/SubChunkPos.h
#	src/mc/world/level/block/registry/BlockTypeRegistry.h
#	src/mc/world/level/levelgen/feature/canyon_feature_utils/CanyonFeatureUtils.h
#	src/mc/world/level/levelgen/structure/BoundingBox.h
#	src/mc/world/level/levelgen/v1/BeardKernel.h
#	src/mc/world/phys/AABB.h
#	src/mc/world/phys/HitResult.h
#	xmake.lua

+++ src/mc/world/actor/animation/BoneBindingMethod.h
@@ -2,7 +2,7 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class BoneBindingMethod : int {
+enum class BoneBindingMethod : uchar {
     Unset                  = 0,
     LocalHierarchy         = 1,
     MatchingParentBoneName = 2,
